requirements.txt
asgiref==3.2.3
Django==2.1.7
django-pyodbc-azure==2.1.0.0
django-rest-framework==0.1.0
django-taggit==0.23.0
djangorestframework==3.10.3
Pillow==6.2.1
pkg-resources==0.0.0
pyodbc==4.0.27
pytz==2019.3
sqlparse==0.3.0
